requirements.txt
altgraph==0.17.4
contourpy==1.2.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.53.1
fpdf==1.7.2
ifcopenshell==0.7.10
isodate==0.6.1
kiwisolver==1.4.5
lark==1.1.9
matplotlib==3.9.1.post1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pefile==2023.2.7
pillow==10.4.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32-ctypes==0.2.2
scipy==1.14.0
seaborn==0.13.2
setuptools==72.2.0
shapely==2.0.5
six==1.16.0
tkinterdnd2==0.4.2
typing_extensions==4.12.2
tzdata==2024.1
CTkMessagebox==2.7